pyproject.toml
[tool.poetry]
name = "fclip"
version = "0.1.0"
description = ""
authors = ["ToDucThanh <toducthanh155@gmail.com>"]
readme = "README.md"
packages = [{include = "fclip", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
opencv-python = "^4.9.0.80"
torch = [
    {version = "^2.0.0", source = "PyPI", platform = "darwin" },
    {version = "2.0.0", platform = "linux", source = "pytorch-cpu-src"}
]
torchvision = [
    { version = "^0.15.0", source = "PyPI", platform = "darwin" },
    {version = "0.15.1", platform = "linux", source = "pytorch-cpu-src"}
]
pyyaml = "^6.0.1"
docopt = "^0.6.2"
matplotlib = "^3.8.3"
scikit-image = "^0.22.0"

[[tool.poetry.source]]
name = "pytorch-cpu-src"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
pytest = "^8.0.0"
ruff = "^0.2.1"
isort = "^5.13.2"
pre-commit = "^3.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
